Json parse
Searching is very slow.

+++ app/src/main/java/com/padassist/Data/RealmDouble.java
@@ -12,7 +12,7 @@ public class RealmDouble extends RealmObject {
     public RealmDouble() {
     }
 
-    public RealmDouble(long value) {
+    public RealmDouble(double value) {
         this.value = value;
     }
